+++ qoqo/src/circuit.rs
@@ -436,7 +436,7 @@ impl PyObjectProtocol for CircuitWrapper {
     /// Args:
     ///     self: The PragmaGeneralNoiseWrapper object.
     ///     other: The object to compare self to.
-    ///     op: Whether they should be equal or not.
+    ///     op: Type of comparison.
     ///
     /// Returns:
     ///     Whether the two operations compared evaluated to True or False
@@ -472,7 +472,7 @@ impl PyNumberProtocol for CircuitWrapper {
     ///     other (Operation): The Operation object to be added to self.
     ///
     /// Raises:
-    ///     TypeError: Right hand side can not be converted to Operation or Circuit.
+    ///     TypeError: Right hand side cannot be converted to Operation or Circuit.
     fn __iadd__(&'p mut self, other: Py<PyAny>) -> PyResult<()> {
         Python::with_gil(|py| -> PyResult<()> {
             let other_ref = other.as_ref(py);
@@ -484,7 +484,7 @@ impl PyNumberProtocol for CircuitWrapper {
                 Err(_) => {
                     let other = convert_into_circuit(other_ref).map_err(|x| {
                         pyo3::exceptions::PyTypeError::new_err(format!(
-                            "Right hand side can not be converted to Operation or Circuit {:?}",
+                            "Right hand side cannot be converted to Operation or Circuit {:?}",
                             x
                         ))
                     });
@@ -511,7 +511,7 @@ impl PyNumberProtocol for CircuitWrapper {
     ///
     /// Raises:
     ///     TypeError: Left hand side can not be converted to Circuit.
-    ///     TypeError: Right hand side can not be converted to Operation or Circuit.
+    ///     TypeError: Right hand side cannot be converted to Operation or Circuit.
     fn __add__(lhs: Py<PyAny>, rhs: Py<PyAny>) -> PyResult<CircuitWrapper> {
         Python::with_gil(|py| -> PyResult<CircuitWrapper> {
             let (lhs_ref, rhs_ref) = (lhs.as_ref(py), rhs.as_ref(py));
@@ -525,7 +525,7 @@ impl PyNumberProtocol for CircuitWrapper {
                 Err(_) => {
                     let other = convert_into_circuit(rhs_ref).map_err(|_| {
                         pyo3::exceptions::PyTypeError::new_err(
-                            "Right hand side can not be converted to Operation or Circuit",
+                            "Right hand side cannot be converted to Operation or Circuit",
                         )
                     })?;
                     Ok(CircuitWrapper {
@@ -611,10 +611,10 @@ pub fn convert_into_circuit(input: &PyAny) -> Result<Circuit, QoqoError> {
     // compiled python packages are involved
     let get_version = input
         .call_method0("_qoqo_versions")
-        .map_err(|_| QoqoError::CannotExtractCircuit)?;
+        .map_err(|_| QoqoError::CannotExtractObject)?;
     let version = get_version
         .extract::<(&str, &str)>()
-        .map_err(|_| QoqoError::CannotExtractCircuit)?;
+        .map_err(|_| QoqoError::CannotExtractObject)?;
     let mut rsplit = ROQOQO_VERSION.split('.').take(2);
     let mut qsplit = QOQO_VERSION.split('.').take(2);
     let rver = format!(
@@ -631,15 +631,16 @@ pub fn convert_into_circuit(input: &PyAny) -> Result<Circuit, QoqoError> {
     if version == test_version {
         let get_bytes = input
             .call_method0("to_bincode")
-            .map_err(|_| QoqoError::CannotExtractCircuit)?;
+            .map_err(|_| QoqoError::CannotExtractObject)?;
         let bytes = get_bytes
             .extract::<Vec<u8>>()
-            .map_err(|_| QoqoError::CannotExtractCircuit)?;
-        deserialize(&bytes[..]).map_err(|_| QoqoError::CannotExtractCircuit)
+            .map_err(|_| QoqoError::CannotExtractObject)?;
+        deserialize(&bytes[..]).map_err(|_| QoqoError::CannotExtractObject)
     } else {
         Err(QoqoError::VersionMismatch)
     }
 }
+
 /// Iterator for iterating over Operations in a Circuit.
 #[pyclass(name = "OperationIterator", module = "qoqo")]
 #[derive(Debug)]